its kinda working :)

+++ usleep.c
@@ -6,13 +6,13 @@
 /*   By: emohamed <emohamed@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/15 12:16:27 by emohamed          #+#    #+#             */
-/*   Updated: 2023/06/15 19:47:10 by emohamed         ###   ########.fr       */
+/*   Updated: 2023/06/16 12:33:54 by emohamed         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-long int	get_current_t(void)
+long long	get_current_t(void)
 {
 	struct timeval	tv;
 
@@ -22,8 +22,8 @@ long int	get_current_t(void)
 
 void	sleeping(int n)
 {
-	long int	save_time;
-	long int	end;
+	long long	save_time;
+	long long	end;
 
 	save_time = get_current_t();
 	end = n + save_time;
